+++ src/containers/FrameRows.cpp
@@ -185,7 +185,7 @@ void osdf::FrameRows::removeRow(const std::int64_t index) {
 }
 
 void osdf::FrameRows::removeRows(const std::vector<bool> & keepRows) {
-  if (keepRows.size() == data_.getSizeRows()) {
+  if (keepRows.size() == static_cast<std::size_t>(data_.getSizeRows())) {
     for (std::int64_t i = (keepRows.size() - 1); i >= 0; --i) {
       if (!keepRows[i]) {
           removeRow(i);
